compile warning fixed
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@3206 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/WorldSession.h
@@ -239,7 +239,7 @@ public:
 
 	ASCENT_INLINE void QueuePacket(WorldPacket* packet)
 	{
-		m_lastPing = UNIXTIME;
+		m_lastPing = (uint32)UNIXTIME;
 		_recvQueue.Push(packet);
 	}
